added new task sol. for ADS

+++ ADSweek1/recursive/task5.java
@@ -8,7 +8,7 @@ public class task5 {
         int a = scanner.nextInt(), n = scanner.nextInt();
         System.out.println(pow(a,n));
     }
-    public static int pow(int a, int n){
+    public static float pow(int a, int n){
         if (n == 1) return a;
         else if (n == 0) return 1;
         else if (n > 1) return a * pow(a, n - 1);
